Fixed integration tests
