+++ .github/workflows/tests.yml
@@ -71,6 +71,11 @@ jobs:
         run: npm run test:db
         working-directory: application/backend
 
+      # integration tests run only in the backend and concentrate on auth/http/trpc mechanics
+      - name: Run integration tests
+        run: npm run test:integration
+        working-directory: application/backend
+
       # build our React frontend
       - run: npm run build:dev
         working-directory: application/frontend
