+++ application/backend/tests/jest.setup.ts
@@ -1,3 +1,7 @@
 import "reflect-metadata";
 
 jest.setTimeout(15000);
+
+process.env["ELSA_DATA_VERSION"] = "jesttest";
+process.env["ELSA_DATA_BUILT"] = "today";
+process.env["ELSA_DATA_REVISION"] = "123456";
