+++ application/backend/tests/service-tests/audit-log-release.test.ts
@@ -2,7 +2,7 @@ import { AuthenticatedUser } from "../../src/business/authenticated-user";
 import * as edgedb from "edgedb";
 import e from "../../dbschema/edgeql-js";
 import { beforeEachCommon } from "./releases.common";
-import { registerTypes } from "./setup";
+import { registerTypes } from "../test-dependency-injection.common";
 import { AuditLogService } from "../../src/business/services/audit-log-service";
 import { addSeconds } from "date-fns";
