+++ application/backend/tests/service-tests/release.participation.test.ts
@@ -1,7 +1,7 @@
 import { AuthenticatedUser } from "../../src/business/authenticated-user";
 import assert from "assert";
 import { beforeEachCommon } from "./releases.common";
-import { registerTypes } from "./setup";
+import { registerTypes } from "../test-dependency-injection.common";
 import { ReleaseService } from "../../src/business/services/release-service";
 import { Client } from "edgedb";
 import { ReleaseParticipationService } from "../../src/business/services/release-participation-service";
