+++ application/backend/tests/service-tests/releases.application-coding.test.ts
@@ -2,7 +2,7 @@ import { Client, createClient } from "edgedb";
 import { AuthenticatedUser } from "../../src/business/authenticated-user";
 import { findDatabaseRelease } from "./utils";
 import { beforeEachCommon } from "./releases.common";
-import { registerTypes } from "./setup";
+import { registerTypes } from "../test-dependency-injection.common";
 import { ReleaseService } from "../../src/business/services/release-service";
 
 let edgeDbClient: Client;
