+++ application/backend/tests/service-tests/releases.cases.test.ts
@@ -10,7 +10,7 @@ import {
 import { ReleaseCaseType, ReleaseSpecimenType } from "@umccr/elsa-types";
 import { PagedResult } from "../../src/api/helpers/pagination-helpers";
 import { beforeEachCommon } from "./releases.common";
-import { registerTypes } from "./setup";
+import { registerTypes } from "../test-dependency-injection.common";
 import { ReleaseService } from "../../src/business/services/release-service";
 import { Client } from "edgedb";
 import {
